+++ src/europarser/api/static/new_wui.js
@@ -88,110 +88,4 @@ Dropzone.options.myDropzone = {
                 }
             }
 
-            document.getElementById('loader').style.display = "block";
-            document.getElementById('submit-all').style.display = "none";
-
-
-            len_ = document.getElementsByClassName("myInput").length
-            for (let i = 0; i < len_; i++) {
-                document.getElementsByClassName("myInput")[i].disabled = true;
-                document.getElementsByClassName("myInput")[i].cursor = "not-allowed";
-                document.getElementsByClassName("myDivInput")[i].style.opacity = "0.5";
-            }
-            alert(1)
-            xhr.responseType = 'blob';
-            xhr.onload = function (e) {
-                alert(1)
-                if (e.currentTarget.status > 300) {
-                    document.getElementById('loader').style.display = "none";
-                    document.getElementById('error').innerHTML = e.currentTarget.statusText;
-                    document.getElementById('error').style.display = "block";
-                }
-                var blob = e.currentTarget.response;
-                var contentDispo = e.currentTarget.getResponseHeader('Content-Disposition');
-                // https://stackoverflow.com/a/23054920/
-                var fileName = contentDispo.match(/filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/)[1];
-                document.getElementById('download').style.display = "block";
-                document.getElementById('loader').style.display = "none";
-                saveBlob(blob, fileName);
-            }
-            xhr.send(formData)
-        });
-    }
-}
-
-// Popper.js
-const tooltip_triggers = document.querySelectorAll(".tooltip-trigger");
-
-tooltip_triggers.forEach((trigger) => {
-        const parent = trigger.parentElement;
-        const label = trigger.getAttribute('aria-label');
-        const tooltip_ = document.querySelector("#" + label);
-
-        const popperInstance = window.Popper.createPopper(trigger, tooltip_, {
-            placement: 'right',
-            modifiers: [
-                {
-                    name: 'offset',
-                    options: {
-                        offset: [0, 20],
-                    },
-                },
-                {
-                    name: 'flip',
-                    options: {
-                        fallbackPlacements: ['left', 'bottom'],
-                    },
-                },
-                {
-                    name: "boundary",
-
-                }
-            ],
-        });
-        const show = () => {
-            // Make the tooltip visible
-            tooltip_.setAttribute('data-show', '');
-
-            // Enable the event listeners
-            popperInstance.setOptions((options) => ({
-                ...options,
-                modifiers: [
-                    ...options.modifiers,
-                    {name: 'eventListeners', enabled: true},
-                ],
-            }));
-
-            // Update its position
-            popperInstance.update();
-        }
-
-        const hide = () => {
-            // Hide the tooltip
-            tooltip_.removeAttribute('data-show');
-
-            // Disable the event listeners
-            popperInstance.setOptions((options) => ({
-                ...options,
-                modifiers: [
-                    ...options.modifiers,
-                    {name: 'eventListeners', enabled: false},
-                ],
-            }));
-        }
-
-        const showEvents = ['mouseenter', 'focus'];
-        const hideEvents = ['mouseleave', 'blur'];
-
-        showEvents.forEach((event) => {
-                parent.addEventListener(event, show);
-            }
-        );
-
-        hideEvents.forEach((event) => {
-                parent.addEventListener(event, hide);
-            }
-        );
-    }
-);
