+++ src/europarser/api/templates/new_wui.html
@@ -27,7 +27,9 @@
     <link href="{{ host + '/static/basic.min.css' }}" rel="stylesheet">
     <link href="{{ host + '/static/new_wui.css' }}" rel="stylesheet">
     <script src="{{ host + '/static/dropzone.min.js' }}"></script>
-    <script type="module" src="{{ host + '/static/new_wui.js' }}"></script>
+    <script src="{{ host + '/static/new_wui.js' }}"></script>
+    <script type="module" src="{{ host + '/static/popover.js' }}"></script>
+
     {#    <script src="{{ host + '/static/jquery-3.6.0.min.js' }}"></script>#}
     {#    <script src="{{ host + '/static/bootstrap.min.js' }}"></script>#}
 </head>
